pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "enterprise-rag"
version = "0.1.0"
description = "Enterprise RAG System with LangChain"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain>=0.3.12",
    "langchain-core>=0.3.27",
    "langchain-community>=0.3.12",
    "langchain-openai>=0.2.14",
    "langchain-text-splitters>=0.3.4",
    "chromadb>=0.5.23",
    "python-dotenv>=1.0.1",
    "pydantic>=2.10.3",
    "loguru>=0.7.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

